Adds Screen Class
Breaks out visual stuff into a screen class.

There is an error in the screen class that needs fixing-- the
default screen does not load, but the rest load when an event is
fired. Hmm.

The screen class just cycles through some images which are lifted
entirely from the lazyfoo tutorial for SDL2 (and are copyrighted)
and will be replaced in the final version of this app.