Migrates code to sublime
This repo is meant to be educational for me. CLion generating files
for me was starting to make it difficult to really be intentional
about the code I am writing. I want to practice in a super light-
weight environment. I may regret this. I have already struggled
without breakpoints. I may revert this. For now, we trudge on.