Migrates project to visual studio
Hopefully this is the final move. I did this to allow for a more
reasonable setup and debugging environment. I've included a sandbox
environment as well as the dll library for zengarden.

I need to automate the clean/rebuild/replace zengarden.dll in
sandbox because that is VERY annoying to do manually every time I
reload. I also need to add a relative path to the images folder
but I was struggling to make it work so it is what it is.

I have renamed Game to Application to be more general.